made IndexProvider configurable as a Spring bean
