+++ spring-data-neo4j/src/main/java/org/springframework/data/neo4j/support/Neo4jTemplate.java
@@ -16,9 +16,22 @@
 
 package org.springframework.data.neo4j.support;
 
+import static org.springframework.data.neo4j.support.ParameterCheck.*;
+
+import java.util.Map;
+
+import javax.annotation.PostConstruct;
+import javax.validation.Validator;
+
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.neo4j.graphdb.*;
+import org.neo4j.graphdb.DynamicRelationshipType;
+import org.neo4j.graphdb.GraphDatabaseService;
+import org.neo4j.graphdb.Node;
+import org.neo4j.graphdb.Path;
+import org.neo4j.graphdb.PropertyContainer;
+import org.neo4j.graphdb.Relationship;
+import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.index.Index;
 import org.neo4j.graphdb.traversal.TraversalDescription;
 import org.neo4j.helpers.collection.ClosableIterable;
@@ -56,12 +69,6 @@ import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.TransactionCallback;
 import org.springframework.transaction.support.TransactionTemplate;
-
-import javax.annotation.PostConstruct;
-import javax.validation.Validator;
-import java.util.Map;
-
-import static org.springframework.data.neo4j.support.ParameterCheck.notNull;
 
 /**
  * Mediator class for the graph related services like the {@link GraphDatabaseService}, the used
@@ -546,10 +553,12 @@ public class Neo4jTemplate implements Neo4jOperations, EntityPersister {
         }
     }
 
+    @Override
     public <T extends PropertyContainer> Index<T> getIndex(String indexName, Class<?> indexedType) {
         return getIndexProvider().getIndex(indexedType, indexName);
     }
 
+    @Override
     public <T extends PropertyContainer> Index<T> getIndex(Class<?> indexedType, String propertyName) {
         final Neo4jPersistentEntityImpl<?> persistentEntity = getPersistentEntity(indexedType);
         final Neo4jPersistentProperty property = persistentEntity.getPersistentProperty(propertyName);
